+++ test/runtests.jl
@@ -21,6 +21,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_coeftable.jl")
         include("unit/test_sunny.jl")
         include("unit/test_introspect.jl")
+        include("unit/test_threading.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")